+++ frontend/src/pages/config/entities/entity-ui-mapping.tsx
@@ -1,7 +1,7 @@
 import { ComponentType, ReactNode } from 'react'
 import { EntityName, EntityType, MasterDataMaps } from '@vlight/types'
 import { css } from '@linaria/core'
-import { mapFixtureList } from '@vlight/controls'
+import { FixtureMappingPrefix, mapFixtureList } from '@vlight/controls'
 
 import {
   iconFixtureType,
@@ -15,6 +15,7 @@ import { FixtureTypeMapShape } from '../../../widgets/map/map-shape'
 import { masterDataMaps } from '../../../api/masterdata'
 import { apiState } from '../../../api/api-state'
 import { getOccupiedFixtureChannels } from '../../../util/fixtures'
+import { editEntity, setLiveChaseState, setLiveMemoryState } from '../../../api'
 
 import { FixtureTypeEditor } from './editors/fixture-type-editor'
 import { FixtureEditor } from './editors/fixture-editor'
@@ -59,6 +60,9 @@ export interface EntityEntry<T extends EntityName> {
 
   /** Function to create a new entry with default values. */
   newEntityFactory?: () => Omit<EntityType<T>, 'id'>
+
+  /** Callback when an entry is deleted. */
+  onDelete?: (entry: EntityType<T>) => void
 }
 
 /**
@@ -129,6 +133,54 @@ export const entityUiMapping: { [key in EntityName]?: EntityEntry<key> } = {
         </div>
       </>
     ),
+    onDelete: group => {
+      // replace group mapping string with fixture strings in chases and memories
+
+      const { fixtures } = group
+      const groupMappingString = FixtureMappingPrefix.Group + group.id
+      const replaceGroupInMapping = (mapping: string[]) =>
+        mapping.includes(groupMappingString)
+          ? mapping.flatMap(mappingString =>
+              mappingString === groupMappingString ? fixtures : mappingString
+            )
+          : mapping
+
+      const { memories } = apiState.rawMasterData!
+
+      for (const memory of memories) {
+        if (
+          memory.scenes.some(
+            scene => replaceGroupInMapping(scene.members) !== scene.members
+          )
+        ) {
+          editEntity('memories', {
+            ...memory,
+            scenes: memory.scenes.map(scene => ({
+              ...scene,
+              members: replaceGroupInMapping(scene.members),
+            })),
+          })
+        }
+      }
+
+      const { liveChases, liveMemories } = apiState
+
+      const liveEntityUpdates = [
+        [liveMemories, setLiveMemoryState],
+        [liveChases, setLiveChaseState],
+      ] as const
+      for (const [entries, updateFn] of liveEntityUpdates) {
+        for (const [id, entry] of Object.entries(entries)) {
+          if (entry.members.includes(groupMappingString)) {
+            updateFn(
+              id,
+              { members: replaceGroupInMapping(entry.members) },
+              true
+            )
+          }
+        }
+      }
+    },
   },
   memories: {
     name: 'Memories',