+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Improvements to the configuration
   - Added button to clone a configuration entry
   - Added ability to position fixtures on the map via mouse and touch
+  - When deleting a group, members chases and memories are changed to contain the group's members instead
 - Added fixture dialogs to map page
 - Added corner button to toggle fullscreen mode
 - Fixed warning when changing the configuration